move prim/scal/meta to prim/meta
