+++ stan/math/fwd/meta/is_fvar.hpp
@@ -2,7 +2,7 @@
 #define STAN_MATH_FWD_META_IS_FVAR_HPP
 
 #include <stan/math/fwd/core.hpp>
-#include <stan/math/prim/scal/meta/is_fvar.hpp>
+#include <stan/math/prim/meta/is_fvar.hpp>
 #include <type_traits>
 
 namespace stan {
