+++ stan/math/fwd/meta/partials_type.hpp
@@ -3,7 +3,7 @@
 
 #include <stan/math/fwd/core.hpp>
 #include <stan/math/fwd/meta/is_fvar.hpp>
-#include <stan/math/prim/scal/meta/partials_type.hpp>
+#include <stan/math/prim/meta/partials_type.hpp>
 #include <type_traits>
 
 namespace stan {
