+++ stan/math/mix/meta.hpp
@@ -1,7 +1,7 @@
 #ifndef STAN_MATH_MIX_META_HPP
 #define STAN_MATH_MIX_META_HPP
 
-#include <stan/math/prim/scal/meta/ad_promotable.hpp>
+#include <stan/math/prim/meta/ad_promotable.hpp>
 
 #include <stan/math/fwd/core.hpp>
 #include <stan/math/fwd/meta/is_fvar.hpp>
