+++ stan/math/prim/meta/VectorBuilderHelper.hpp
@@ -1,7 +1,7 @@
 #ifndef STAN_MATH_PRIM_SCAL_META_VECTORBUILDER_HELPER_HPP
 #define STAN_MATH_PRIM_SCAL_META_VECTORBUILDER_HELPER_HPP
 
-#include <stan/math/prim/scal/meta/VectorBuilderHelper.hpp>
+#include <stan/math/prim/meta/VectorBuilderHelper.hpp>
 #include <stdexcept>
 #include <vector>
