+++ stan/math/prim/meta/ad_promotable.hpp
@@ -1,7 +1,7 @@
 #ifndef STAN_MATH_PRIM_SCAL_META_AD_PROMOTABLE_HPP
 #define STAN_MATH_PRIM_SCAL_META_AD_PROMOTABLE_HPP
 
-#include <stan/math/prim/scal/meta/bool_constant.hpp>
+#include <stan/math/prim/meta/bool_constant.hpp>
 #include <type_traits>
 
 namespace stan {
