+++ stan/math/prim/meta/append_return_type.hpp
@@ -2,7 +2,7 @@
 #define STAN_MATH_PRIM_MAT_META_APPEND_RETURN_TYPE_HPP
 
 #include <stan/math/prim/mat/fun/Eigen.hpp>
-#include <stan/math/prim/scal/meta/return_type.hpp>
+#include <stan/math/prim/meta/return_type.hpp>
 #include <vector>
 
 namespace stan {
