+++ stan/math/prim/meta/as_array_or_scalar.hpp
