+++ stan/math/prim/meta/as_scalar.hpp
