+++ stan/math/prim/meta/broadcast_array.hpp
@@ -1,7 +1,7 @@
 #ifndef STAN_MATH_PRIM_SCAL_META_BROADCAST_ARRAY_HPP
 #define STAN_MATH_PRIM_SCAL_META_BROADCAST_ARRAY_HPP
 
-#include <stan/math/prim/scal/meta/broadcast_array.hpp>
+#include <stan/math/prim/meta/broadcast_array.hpp>
 #include <stan/math/prim/mat/fun/Eigen.hpp>
 #include <stdexcept>
