+++ stan/math/prim/meta/child_type.hpp
