+++ stan/math/prim/meta/contains_fvar.hpp
@@ -1,9 +1,9 @@
 #ifndef STAN_MATH_PRIM_SCAL_META_CONTAINS_FVAR_HPP
 #define STAN_MATH_PRIM_SCAL_META_CONTAINS_FVAR_HPP
 
-#include <stan/math/prim/scal/meta/is_fvar.hpp>
-#include <stan/math/prim/scal/meta/scalar_type.hpp>
-#include <stan/math/prim/scal/meta/disjunction.hpp>
+#include <stan/math/prim/meta/is_fvar.hpp>
+#include <stan/math/prim/meta/scalar_type.hpp>
+#include <stan/math/prim/meta/disjunction.hpp>
 
 namespace stan {
