+++ stan/math/prim/meta/contains_std_vector.hpp
@@ -1,9 +1,9 @@
 #ifndef STAN_MATH_PRIM_SCAL_META_CONTAINS_STD_VECTOR_HPP
 #define STAN_MATH_PRIM_SCAL_META_CONTAINS_STD_VECTOR_HPP
 
-#include <stan/math/prim/scal/meta/bool_constant.hpp>
-#include <stan/math/prim/scal/meta/disjunction.hpp>
-#include <stan/math/prim/scal/meta/is_vector.hpp>
+#include <stan/math/prim/meta/bool_constant.hpp>
+#include <stan/math/prim/meta/disjunction.hpp>
+#include <stan/math/prim/meta/is_vector.hpp>
 #include <type_traits>
 
 namespace stan {
