+++ stan/math/prim/meta/contains_vector.hpp
@@ -1,8 +1,8 @@
 #ifndef STAN_MATH_PRIM_SCAL_META_CONTAINS_VECTOR_HPP
 #define STAN_MATH_PRIM_SCAL_META_CONTAINS_VECTOR_HPP
 
-#include <stan/math/prim/scal/meta/is_vector.hpp>
-#include <stan/math/prim/scal/meta/disjunction.hpp>
+#include <stan/math/prim/meta/is_vector.hpp>
+#include <stan/math/prim/meta/disjunction.hpp>
 
 namespace stan {
 /** \ingroup type_trait
