+++ stan/math/prim/meta/include_summand.hpp
@@ -1,9 +1,9 @@
 #ifndef STAN_MATH_PRIM_SCAL_META_INCLUDE_SUMMAND_HPP
 #define STAN_MATH_PRIM_SCAL_META_INCLUDE_SUMMAND_HPP
 
-#include <stan/math/prim/scal/meta/bool_constant.hpp>
-#include <stan/math/prim/scal/meta/is_constant.hpp>
-#include <stan/math/prim/scal/meta/scalar_type.hpp>
+#include <stan/math/prim/meta/bool_constant.hpp>
+#include <stan/math/prim/meta/is_constant.hpp>
+#include <stan/math/prim/meta/scalar_type.hpp>
 #include <type_traits>
 
 namespace stan {
