+++ stan/math/prim/meta/index_type.hpp
@@ -2,11 +2,11 @@
 #define STAN_MATH_PRIM_SCAL_META_INDEX_TYPE_HPP
 
 #include <stan/math/prim/mat/fun/Eigen.hpp>
-#include <stan/math/prim/scal/meta/index_type.hpp>
-#include <stan/math/prim/scal/meta/require_generics.hpp>
-#include <stan/math/prim/scal/meta/is_eigen.hpp>
-#include <stan/math/prim/scal/meta/index_type.hpp>
-#include <stan/math/prim/scal/meta/require_generics.hpp>
+#include <stan/math/prim/meta/index_type.hpp>
+#include <stan/math/prim/meta/require_generics.hpp>
+#include <stan/math/prim/meta/is_eigen.hpp>
+#include <stan/math/prim/meta/index_type.hpp>
+#include <stan/math/prim/meta/require_generics.hpp>
 #include <type_traits>
 #include <type_traits>
 #include <vector>
