+++ stan/math/prim/meta/is_eigen.hpp
@@ -2,8 +2,8 @@
 #define STAN_MATH_PRIM_SCAL_META_IS_EIGEN_HPP
 
 #include <stan/math/prim/mat/fun/Eigen.hpp>
-#include <stan/math/prim/scal/meta/is_eigen.hpp>
-#include <stan/math/prim/scal/meta/disjunction.hpp>
+#include <stan/math/prim/meta/is_eigen.hpp>
+#include <stan/math/prim/meta/disjunction.hpp>
 #include <type_traits>
 
 namespace stan {
