+++ stan/math/prim/meta/is_var.hpp
