+++ stan/math/prim/meta/is_var_or_arithmetic.hpp
@@ -1,10 +1,10 @@
 #ifndef STAN_MATH_PRIM_SCAL_META_IS_VAR_OR_ARITHMETIC_HPP
 #define STAN_MATH_PRIM_SCAL_META_IS_VAR_OR_ARITHMETIC_HPP
 
-#include <stan/math/prim/scal/meta/bool_constant.hpp>
-#include <stan/math/prim/scal/meta/is_var.hpp>
-#include <stan/math/prim/scal/meta/scalar_type.hpp>
-#include <stan/math/prim/scal/meta/conjunction.hpp>
+#include <stan/math/prim/meta/bool_constant.hpp>
+#include <stan/math/prim/meta/is_var.hpp>
+#include <stan/math/prim/meta/scalar_type.hpp>
+#include <stan/math/prim/meta/conjunction.hpp>
 #include <type_traits>
 
 namespace stan {
