+++ stan/math/prim/meta/is_vector_like.hpp
@@ -1,9 +1,9 @@
 #ifndef STAN_MATH_PRIM_SCAL_META_IS_VECTOR_LIKE_HPP
 #define STAN_MATH_PRIM_SCAL_META_IS_VECTOR_LIKE_HPP
 
-#include <stan/math/prim/scal/meta/bool_constant.hpp>
-#include <stan/math/prim/scal/meta/is_eigen.hpp>
-#include <stan/math/prim/scal/meta/is_vector.hpp>
+#include <stan/math/prim/meta/bool_constant.hpp>
+#include <stan/math/prim/meta/is_eigen.hpp>
+#include <stan/math/prim/meta/is_vector.hpp>
 #include <type_traits>
 
 namespace stan {
