+++ stan/math/prim/meta/max_size.hpp
@@ -1,7 +1,7 @@
 #ifndef STAN_MATH_PRIM_SCAL_META_MAX_SIZE_HPP
 #define STAN_MATH_PRIM_SCAL_META_MAX_SIZE_HPP
 
-#include <stan/math/prim/scal/meta/length.hpp>
+#include <stan/math/prim/meta/length.hpp>
 
 namespace stan {
