+++ stan/math/prim/meta/max_size_mvt.hpp
@@ -1,7 +1,7 @@
 #ifndef STAN_MATH_PRIM_SCAL_META_MAX_SIZE_MVT_HPP
 #define STAN_MATH_PRIM_SCAL_META_MAX_SIZE_MVT_HPP
 
-#include <stan/math/prim/scal/meta/length_mvt.hpp>
+#include <stan/math/prim/meta/length_mvt.hpp>
 #include <cstdlib>
 
 namespace stan {
