+++ stan/math/prim/meta/operands_and_partials.hpp
@@ -2,9 +2,9 @@
 #define STAN_MATH_PRIM_SCAL_META_OPERANDS_AND_PARTIALS_HPP
 
 #include <stan/math/prim/mat/fun/Eigen.hpp>
-#include <stan/math/prim/scal/meta/broadcast_array.hpp>
-#include <stan/math/prim/scal/meta/return_type.hpp>
-#include <stan/math/prim/scal/meta/operands_and_partials.hpp>
+#include <stan/math/prim/meta/broadcast_array.hpp>
+#include <stan/math/prim/meta/return_type.hpp>
+#include <stan/math/prim/meta/operands_and_partials.hpp>
 #include <vector>
 
 namespace stan {
