+++ stan/math/prim/meta/return_type.hpp
@@ -1,8 +1,8 @@
 #ifndef STAN_MATH_PRIM_SCAL_META_RETURN_TYPE_HPP
 #define STAN_MATH_PRIM_SCAL_META_RETURN_TYPE_HPP
 
-#include <stan/math/prim/scal/meta/promote_args.hpp>
-#include <stan/math/prim/scal/meta/scalar_type.hpp>
+#include <stan/math/prim/meta/promote_args.hpp>
+#include <stan/math/prim/meta/scalar_type.hpp>
 
 namespace stan {
