+++ stan/math/prim/meta/scalar_seq_view.hpp
@@ -1,8 +1,8 @@
 #ifndef STAN_MATH_PRIM_SCAL_META_SCALAR_SEQ_VIEW_HPP
 #define STAN_MATH_PRIM_SCAL_META_SCALAR_SEQ_VIEW_HPP
 
-#include <stan/math/prim/scal/meta/scalar_type.hpp>
-#include <stan/math/prim/scal/meta/is_vector_like.hpp>
+#include <stan/math/prim/meta/scalar_type.hpp>
+#include <stan/math/prim/meta/is_vector_like.hpp>
 #include <type_traits>
 
 namespace stan {
