+++ stan/math/prim/meta/size_of.hpp
@@ -1,7 +1,7 @@
 #ifndef STAN_MATH_PRIM_SCAL_META_SIZE_OF_HPP
 #define STAN_MATH_PRIM_SCAL_META_SIZE_OF_HPP
 
-#include <stan/math/prim/scal/meta/is_vector.hpp>
+#include <stan/math/prim/meta/is_vector.hpp>
 #include <cstddef>
 
 namespace stan {
