+++ stan/math/rev/meta/is_var.hpp
@@ -1,7 +1,7 @@
 #ifndef STAN_MATH_REV_META_IS_VAR_HPP
 #define STAN_MATH_REV_META_IS_VAR_HPP
 
-#include <stan/math/prim/scal/meta/is_var.hpp>
+#include <stan/math/prim/meta/is_var.hpp>
 #include <stan/math/rev/core.hpp>
 #include <type_traits>
