+++ stan/math/rev/meta/partials_type.hpp
@@ -2,7 +2,7 @@
 #define STAN_MATH_REV_META_PARTIALS_TYPE_HPP
 
 #include <stan/math/rev/core.hpp>
-#include <stan/math/prim/scal/meta/partials_type.hpp>
+#include <stan/math/prim/meta/partials_type.hpp>
 #include <stan/math/rev/meta/is_var.hpp>
 #include <type_traits>
