+++ test/unit/math/rev/mat/prob/categorical_logit_glm_lpmf_test.cpp
@@ -2,7 +2,7 @@
 #include <stan/math.hpp>
 #include <gtest/gtest.h>
 #include <stan/math/prim/scal/fun/value_of.hpp>
-#include <stan/math/prim/scal/meta/return_type.hpp>
+#include <stan/math/prim/meta/return_type.hpp>
 #include <stan/math/prim/scal/err/check_consistent_size.hpp>
 #include <stan/math/prim/mat/prob/categorical_logit_lpmf.hpp>
 #include <vector>
