+++ test/unit/math/rev/meta/is_var_or_arithmetic_test.cpp
@@ -1,5 +1,5 @@
 #include <stan/math/rev/meta.hpp>
-#include <stan/math/prim/scal/meta/is_var_or_arithmetic.hpp>
+#include <stan/math/prim/meta/is_var_or_arithmetic.hpp>
 #include <gtest/gtest.h>
 #include <vector>
